# NothingWebApi/Dtos/NothingModelDto.cs
using System.Text.Json.Serialization;

namespace NothingWebApi.Dtos;

/// <summary>
/// Dto модели
/// </summary>
public class NothingModelDto
{
    /// <summary>
    /// Идентификатор модели
    /// </summary>
    [JsonPropertyName("id")]
    public required int Id { get; init; }

    /// <summary>
    /// Имя модели
    /// </summary>
    [JsonPropertyName("name")]
    public required string Name { get; init; }
}