# NothingWebApi/openapi_v1.0.yaml
openapi: 3.0.1
info:
  title: NothingWebApi
  version: v1.0
paths:
  /:
    get:
      tags:
        - NothingWebApi
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /NothingWebApi:
    get:
      tags:
        - NothingWebApi
      summary: Получить список моделей
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NothingModelDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NothingModelDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NothingModelDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - NothingWebApi
      summary: Создать новую модель
      description: " Пример запроса:\r\n\r\n     {\r\n\t        \"name\": \"Test\"\r\n     }"
      requestBody:
        description: Данные для создания модели
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNothingModelDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateNothingModelDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateNothingModelDto'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
            application/json:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
            text/json:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - NothingWebApi
      summary: Обновить существующую модель
      description: " Пример запроса:\r\n\r\n     {\r\n         \"id\": 1,\r\n\t        \"name\": \"Test\"\r\n     }"
      requestBody:
        description: Данные для обновления модели
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNothingModelDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateNothingModelDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateNothingModelDto'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
            application/json:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
            text/json:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/NothingWebApi/{id}':
    get:
      tags:
        - NothingWebApi
      summary: Получить модель с указанным идентификатором
      parameters:
        - name: id
          in: path
          description: Идентификатор модели
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
            application/json:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
            text/json:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      tags:
        - NothingWebApi
      summary: Удалить модель с указанным идентификатором
      parameters:
        - name: id
          in: path
          description: Идентификатор графа
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
            application/json:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
            text/json:
              schema:
                $ref: '#/components/schemas/NothingModelDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
    CreateNothingModelDto:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          description: Имя модели
          nullable: true
      additionalProperties: false
      description: Данные для создания модели
    NothingModelDto:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор модели
          format: int32
        name:
          type: string
          description: Имя модели
          nullable: true
      additionalProperties: false
      description: Dto модели
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    UpdateNothingModelDto:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор модели
          format: int32
        name:
          type: string
          description: Имя модели
          nullable: true
      additionalProperties: false
      description: Данные для обновления модели